Moved MainActivity and other files to Shared.
