+++ Android/CameraSample/app/src/main/AndroidManifest.xml
@@ -8,7 +8,7 @@
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
 
     <application
-        android:name="com.bentley.sample.camera.CameraSample"
+        android:name="com.bentley.sample.camera.CameraApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -18,7 +18,7 @@
         tools:targetApi="31"
         android:networkSecurityConfig="@xml/network_security_config">
         <activity
-            android:name="com.bentley.sample.camera.MainActivity"
+            android:name="com.bentley.sample.shared.MainActivity"
             android:configChanges="orientation|screenSize|keyboard|keyboardHidden|layoutDirection|locale|fontScale|uiMode"
             android:exported="true">
             <intent-filter>
