+++ Android/Shared/build.gradle
@@ -33,6 +33,8 @@ dependencies {
 
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.5.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
+    implementation "androidx.lifecycle:lifecycle-runtime-ktx:2.5.1"
     implementation 'com.eclipsesource.minimal-json:minimal-json:0.9.5'
     implementation 'com.google.android.material:material:1.6.1'
     releaseImplementation 'com.github.itwin:mobile-sdk-android:0.10.30'
