+++ Android/Shared/src/main/java/com/bentley/sample/shared/FileHelper.kt
@@ -12,7 +12,16 @@ import java.io.File
 import java.io.FileOutputStream
 import java.io.InputStream
 
+/**
+ * File related utility functions.
+ */
 object FileHelper {
+    /**
+     * Determines the display name for the uri using the content resolver.
+     * @param uri The input Uri.
+     * @param contentResolver The content resolver to query for the display name.
+     * @return The display name or null if it wasn't found.
+     */
     fun getFileDisplayName(uri: Uri, contentResolver: ContentResolver): String? {
         contentResolver.query(uri, null, null, null, null, null)?.let { cursor ->
             if (cursor.moveToFirst()) {
@@ -26,11 +35,19 @@ object FileHelper {
         return null
     }
 
-    private fun copyFile(inputStream: InputStream, dir: File, displayName: String): String {
-        if (!dir.exists()) {
-            dir.mkdirs()
+    /**
+     * Copies the input stream to the destination directory using the input display name.
+     * @param inputStream The stream to copy.
+     * @param destDir The destination directory.
+     * @param displayName The file name.
+     * @return The full path of the output file.
+     */
+    private fun copyFile(inputStream: InputStream, destDir: File, displayName: String): String {
+        if (!destDir.exists()) {
+            destDir.mkdirs()
         }
-        val dstPath = dir.absolutePath + "/" + displayName
+        // @TODO: use File to combine the path and the dir instead of hardcoding the /
+        val dstPath = destDir.absolutePath + "/" + displayName
         val outputStream = FileOutputStream(dstPath)
         val buffer = ByteArray(1024)
         var length: Int
@@ -42,6 +59,13 @@ object FileHelper {
         return dstPath
     }
 
+    /**
+     * Copies the input uri to the destination directory using the display name.
+     * @param context The context for getting external files and the content resolver.
+     * @param uri The input Uri to copy.
+     * @param destDir The destination directory.
+     * @return The full path of the copied file, null if it failed.
+     */
     fun copyToExternalFiles(context: Context, uri: Uri, destDir: String, displayName: String): String? {
         var result: String? = null
         context.getExternalFilesDir(null)?.let { filesDir ->
@@ -53,10 +77,23 @@ object FileHelper {
         return result
     }
 
+    /**
+     * Gets the external files in the input directory name.
+     * @param context The context.
+     * @param dirName The directory name.
+     * @return A list of external files, possibly empty.
+     */
     fun getExternalFiles(context: Context, dirName: String): List<String> {
         return context.getExternalFilesDir(dirName)?.listFiles()?.map { it.toString() } ?: emptyList()
     }
 
+    /**
+     * Gets the external files in the input directory name that end with the input extension.
+     * @param context The context.
+     * @param dirName The directory name.
+     * @param extension The file extension to filter with.
+     * @return A list of external files, possibly empty.
+     */
     fun getExternalFiles(context: Context, dirName: String, extension: String): List<String> {
         return getExternalFiles(context, dirName).filter { name ->
             extension.isEmpty() || name.endsWith(extension, true)
