+++ Android/Shared/src/main/res/layout/activity_main.xml
